+++ trepanxpy/__init__.py
@@ -1,11 +1,8 @@
 """
 trepan debugger for x-python
 
-Copyright 2020 Rocky Bernstein
+Copyright 2020-2021 Rocky Bernstein
 """
 __docformat__ = "restructuredtext"
 
-from trepanxpy.version import VERSION
-
-# This ensures VERSION will appear in pydoc
-__version__ = VERSION
+from trepanxpy.version import __version__  # noqa