+++ NEWS.md
@@ -1,3 +1,17 @@
+1.1.1 2024-07-22
+================
+
+* Adjust for newer xdis, x-python, and trepan3k
+
+1.1.0 2021-11-07
+================
+
+* Adjust for newer xdis 6.0.3, and x-python
+* Add commands:
+  - `vmstack {peek|push|pop}`
+  - `set autostack`
+
+
 1.0.3 2020-06-28
 ================